ModBot/bot/twitch/requirements.txt
twitchio==2.8.2
aiohttp==3.9.5
python-dotenv==1.0.1
